Move useToast to own file
